pyproject.toml
[project]
name = "scheduler"
version = "0.1.0"
description = "Software to schedule jobs to deploy through ci/cd pipelines"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.40",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.10",
    "taskipy>=1.14.1",
]

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'

[tool.ruff]
line-length = 88
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style =  'single'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
# run = 'fastapi dev'
pre_test = 'task lint; task format'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'
